+++ sw/airborne/firmwares/wind_tunnel/main_mb.c
@@ -89,7 +89,7 @@ void dl_parse_msg(void) {
 
   case DL_SETTING : {
     uint8_t i = DL_SETTING_index(dl_buffer);
-    float var = DL_SETTING_value(dl_buffer);
+    double var = DL_SETTING_value(dl_buffer);
     DlSetting(i, var);
     DOWNLINK_SEND_DL_VALUE(&i, &var);
     break;
